Fix workout media upload and detail experience
