+++ AscendApp/Shared/Components/FullScreenPhotoView.swift
@@ -29,7 +29,7 @@ struct FullScreenPhotoView: View {
             // Video Player for videos
             if photo.isVideo {
                 if let player = player {
-                    VideoPlayer(player: player)
+                    FullScreenVideoPlayer(player: player)
                         .ignoresSafeArea()
                         .onAppear {
                             player.play()
@@ -180,8 +180,7 @@ struct FullScreenPhotoView: View {
     
     @ViewBuilder
     private var overlayViews: some View {
-        // Close button
-        VStack {
+        VStack(spacing: 0) {
             HStack {
                 Spacer()
 
@@ -190,50 +189,23 @@ struct FullScreenPhotoView: View {
                 } label: {
                     Image(systemName: "xmark.circle.fill")
                         .font(.system(size: 32))
-                        .foregroundStyle(.white.opacity(0.8))
-                        .background(Color.black.opacity(0.5))
+                        .foregroundStyle(.white.opacity(0.85))
+                        .background(Color.black.opacity(0.45))
                         .clipShape(Circle())
                 }
             }
-            .padding()
+            .padding(.horizontal, 16)
+            .padding(.top, 12)
 
-            Spacer()
+            Spacer(minLength: 0)
         }
-
-        // Media info overlay (bottom)
-        VStack {
-            Spacer()
-
-            HStack {
-                VStack(alignment: .leading, spacing: 4) {
-                    Text("Uploaded \(photo.uploadedAt.formatted(.dateTime.month().day().hour().minute()))")
-                        .font(.caption)
-                        .foregroundStyle(.white.opacity(0.8))
-                    
-                    if photo.isVideo, let duration = photo.duration {
-                        Text("Duration: \(formatDuration(duration))")
-                            .font(.caption)
-                            .foregroundStyle(.white.opacity(0.8))
-                    }
-                }
-
-                Spacer()
-            }
-            .padding()
-            .background(
-                LinearGradient(
-                    colors: [.clear, .black.opacity(0.6)],
-                    startPoint: .top,
-                    endPoint: .bottom
-                )
-            )
-        }
+        .safeAreaPadding(.top)
     }
     
     private func loadVideo() async {
         let asset = AVURLAsset(url: photo.url)
         // Preload to avoid main thread blocking
-        try? await asset.load(.isPlayable)
+        _ = try? await asset.load(.isPlayable)
 
         let playerItem = AVPlayerItem(asset: asset)
 
@@ -286,12 +258,23 @@ struct FullScreenPhotoView: View {
         offset.width = max(-maxOffsetX, min(maxOffsetX, offset.width))
         offset.height = max(-maxOffsetY, min(maxOffsetY, offset.height))
     }
-    
-    private func formatDuration(_ duration: TimeInterval) -> String {
-        let totalSeconds = Int(duration.rounded())
-        let minutes = totalSeconds / 60
-        let seconds = totalSeconds % 60
-        return "\(minutes):\(seconds < 10 ? "0" : "")\(seconds)"
+}
+
+private struct FullScreenVideoPlayer: UIViewControllerRepresentable {
+    let player: AVPlayer
+
+    func makeUIViewController(context: Context) -> AVPlayerViewController {
+        let controller = AVPlayerViewController()
+        controller.player = player
+        controller.showsPlaybackControls = true
+        controller.allowsPictureInPicturePlayback = false
+        controller.canStartPictureInPictureAutomaticallyFromInline = false
+        controller.updatesNowPlayingInfoCenter = false
+        return controller
+    }
+
+    func updateUIViewController(_ uiViewController: AVPlayerViewController, context: Context) {
+        uiViewController.player = player
     }
 }
